Updated to include the correct paths.
