+++ src/stochastic/classes/rsa_2d/nn_exclusion.py
@@ -1,21 +1,26 @@
-""" File that contains the random sequential adsorption of dimers.
+""" File that contains the random sequential adsorption of monomers with nearest
+    neighbor exclusion.
 """
 
 # ------------------------------------------------------------------------------
 # Imports.
 # ------------------------------------------------------------------------------
 
+# Imports: General.
+import copy as cp
+
 # Imports: User-defined.
-from stochastic.interfaces.RSA_1D_interface import RSA1D
-from stochastic.utilities.RSA_parameters import RSA1DParameters
+from stochastic.interfaces.rsa_2d_interface import RSA2D
+from stochastic.utilities.rsa_parameters import RSA2DParameters
 
 # ------------------------------------------------------------------------------
 # Classes.
 # ------------------------------------------------------------------------------
 
 
-class Dimers(RSA1D):
-    """ Class to simulate random sequential adsorption of dimers.
+class NNExclusion(RSA2D):
+    """ Class to simulate random sequential adsorption with nearest neighbor
+        exclusion for a two-dimensional lattice.
 
         Inherited parameters:
 
@@ -65,8 +70,8 @@ class Dimers(RSA1D):
         """
 
         preheader = ",".join([
-            f"RSA Dimer", f"seed={self.seed}", f"repetitions(n)={self.repetitions}",
-            f"maximum time={self.maximum_time}", f"length={self.length}", f"periodic={self.periodic}"
+            f"RSA-2D NN Exclusion", f"seed={self.seed}", f"repetitions(n)={self.repetitions}",
+            f"maximum time={self.maximum_time}", f"dimensions={self.dimensions}", f"periodic={self.periodic}"
         ])
 
         return preheader
@@ -79,38 +84,45 @@ class Dimers(RSA1D):
         """ Tries to perform an adsorption operation; i.e., select a random site
             in the lattice onto which to adsorb.
         """
-        generator = self.random_generator
-        site = generator.integers(0, self.length) if self.periodic else generator.integers(0, self.length - 1)
+
+        generator = self.random_generator.integers
+        site = [generator(0, self.dimensions[i]) for i in range(2)]
+
         self.attempts += 1
+        site = tuple(site)
         if self.validate_adsorb(site):
-            site_ = self.normalize_site(site + 1)
-            self.lattice[site] = RSA1D.OCCUPIED
-            self.lattice[site_] = RSA1D.OCCUPIED
+            self.lattice[site[0]][site[1]] = RSA2D.OCCUPIED
             self.attempts_successful += 1
 
     # --------------------------------------------------------------------------
     # Validate Methods.
     # --------------------------------------------------------------------------
 
-    def validate_adsorb(self, site: int) -> bool:
+    def validate_adsorb(self, site: tuple) -> bool:
         """ Determines if the given site can adsorb a particle.
 
-            :param site: The site to be examined. Must be an integer number.
+            :param site: The tuple that represents the site where the particle
+             is to be adsorbed.
 
             :return: If the site is empty and its inmediate neighbors are empty.
         """
 
-        site_ = self.normalize_site(site)
-        if not self.lattice[site_] == RSA1D.EMPTY:
+        if not self.lattice[site[0]][site[1]] == RSA2D.EMPTY:
             return False
 
-        examined = {site_}
-        if ((site + 1) >= self.length and self.periodic) or (site + 1) < self.length:
-            site_ = self.normalize_site(site + 1)
-            if (site_ not in examined) and not self.lattice[site_] == RSA1D.EMPTY:
-                return False
+        examined = set()
+        indexes = ((1, 0), (-1, 0), (0, 1), (0, -1))
+        for indexes_ in indexes:
+            indexes_ = tuple(index + site[i] for i, index in enumerate(indexes_))
+            indexes_ = self.normalize_site(indexes_)
+            if self.validate_in_lattice(indexes_):
+                examined.add(cp.deepcopy(indexes_))
 
-        return True
+        valid = True
+        for indexes_ in examined:
+            valid = valid and self.lattice[indexes_[0]][indexes_[1]] == RSA2D.EMPTY
+
+        return valid
 
     # $$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$
     # Constructor and Dunder Methods.
@@ -120,10 +132,10 @@ class Dimers(RSA1D):
     # Constructor.
     # --------------------------------------------------------------------------
 
-    def __init__(self, parameters: RSA1DParameters):
+    def __init__(self, parameters: RSA2DParameters):
         """ Initializes the simulation parameters.
 
             :param parameters: A dataclass that contains the adjustable
              parameters of the simulation.
         """
-        super(Dimers, self).__init__(parameters)
+        super(NNExclusion, self).__init__(parameters)
